+++ SDK/robot.h
@@ -3,13 +3,13 @@
 
 class Robot {
     public:
-        explicit Robot(float x, float y):location_x(x), location_y(y) {}
+        explicit Robot(double x, double y):location_x(x), location_y(y) {}
         ~Robot() = default;
 
 
-        void Update(int workstation_id, int carried_item_type, float time_value_coefficient, float collision_value_coefficient, 
-        float angular_velocity, float linear_velocity_x, float linear_velocity_y, float direction,
-        float location_x, float location_y);
+        void Update(int workstation_id, int carried_item_type, double time_value_coefficient, double collision_value_coefficient, 
+        double angular_velocity, double linear_velocity_x, double linear_velocity_y, double direction,
+        double location_x, double location_y);
 
 
 
@@ -19,12 +19,12 @@ class Robot {
         //下面是机器人的属性
         int workstation_id;
         int carried_item_type;
-        float time_value_coefficient;
-        float collision_value_coefficient;
-        float angular_velocity;
-        float linear_velocity_x, linear_velocity_y;
-        float direction;
-        float location_x, location_y;
+        double time_value_coefficient;
+        double collision_value_coefficient;
+        double angular_velocity;
+        double linear_velocity_x, linear_velocity_y;
+        double direction;
+        double location_x, location_y;
 
 };
